refactor(tests): improve readability of test assertions
