--- src/components/BlogPostList/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./BlogPostList";
-export { default as BlogPostList } from "./BlogPostList";
