+++ src/components/Container/__tests__/Container.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import Container from "../Container";
+import Container from "..";
 
 describe("Container", () => {
   it("renders its children with the base and custom classes", () => {
