--- src/components/Container/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Container";
-export { default as Container } from "./Container";
