--- src/components/ContributorsGrid/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ContributorsGrid";
-export { default as ContributorsGrid } from "./ContributorsGrid";
