--- src/components/Divider/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Divider";
-export { default as Divider } from "./Divider";
