+++ src/components/Footer/__tests__/Footer.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import Footer from "../Footer";
+import Footer from "..";
 
 describe("Footer", () => {
   it("renders the footer copy and policy link", () => {
