--- src/components/Footer/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Footer";
-export { default as Footer } from "./Footer";
