--- src/components/Header/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Header";
-export { default as Header } from "./Header";
