--- src/components/HeroSection/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./HeroSection";
-export { default as HeroSection } from "./HeroSection";
