--- src/components/IssueJumpSection/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./IssueJumpSection";
-export { default as IssueJumpSection } from "./IssueJumpSection";
