--- src/components/IssueList/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./IssueList";
-export { default as IssueList } from "./IssueList";
