--- src/components/MeetSection/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./MeetSection";
-export { default as MeetSection } from "./MeetSection";
