+++ src/components/Menu/__tests__/Menu.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import Menu from "../Menu";
+import Menu from "..";
 
 const mockNavigation = vi.hoisted(() => ({
   pathname: "/",
