--- src/components/Menu/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Menu";
-export { default as Menu } from "./Menu";
