--- src/components/OpenSourceSection/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./OpenSourceSection";
-export { default as OpenSourceSection } from "./OpenSourceSection";
