--- src/components/QuotesCarousel/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./QuotesCarousel";
-export { default as QuotesCarousel } from "./QuotesCarousel";
