+++ src/components/QuotesCarousel/index.tsx
