--- src/components/ReposCarousel/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ReposCarousel";
-export { default as ReposCarousel } from "./ReposCarousel";
