--- src/components/RichText/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./RichText";
-export { default as RichText } from "./RichText";
