--- src/components/ShareButtons/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ShareButtons";
-export { default as ShareButtons } from "./ShareButtons";
