--- src/components/SimpleContentPage/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./SimpleContentPage";
-export { default as SimpleContentPage } from "./SimpleContentPage";
