--- src/components/WhatIsHieroSection/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./WhatIsHieroSection";
-export { default as WhatIsHieroSection } from "./WhatIsHieroSection";
