+++ src/components/WhatIsHieroSection/index.tsx
